Main/disease_detection/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
Django==3.2.25
flatbuffers==1.12
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==3.8.0
idna==3.10
importlib-metadata==6.7.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.4.4
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
